convert to Double arithmetic for action costs - required changing nearly every domain - update to the EES family of algorithms work and added new data structures RBTree, Binheap, GEQueue which are used in the EES algorithms

+++ src/main/kotlin/edu/unh/cs/ai/realtimesearch/environment/Domain.kt
@@ -103,5 +103,5 @@ interface Domain<State> {
     fun appendDomainSpecificResults(results: ExperimentResult) {}
 }
 
-data class SuccessorBundle<out State>(val state: State, val action: Action, val actionCost: Long)
+data class SuccessorBundle<out State>(val state: State, val action: Action, val actionCost: Double)
