+++ src/main/kotlin/edu/unh/cs/ai/realtimesearch/planner/realtime/LssLrtaStarPlanner.kt
@@ -227,7 +227,7 @@ class LssLrtaStarPlanner<StateType : State<StateType>>(val domain: Domain<StateT
             }
 
             // Add the current state as the predecessor of the child state
-            successorNode.predecessors.add(Edge(node = sourceNode, action = successor.action, actionCost = successor.actionCost))
+            successorNode.predecessors.add(Edge(node = sourceNode, action = successor.action, actionCost = successor.actionCost.toLong()))
 
             // Skip if we got back to the parent
             if (successorState == sourceNode.parent.state) {
@@ -239,10 +239,10 @@ class LssLrtaStarPlanner<StateType : State<StateType>>(val domain: Domain<StateT
             if (successorNode.cost > successorGValueFromCurrent) {
                 // here we generate a state. We store it's g value and remember how to get here via the treePointers
                 successorNode.apply {
-                    cost = successorGValueFromCurrent
+                    cost = successorGValueFromCurrent.toLong()
                     parent = sourceNode
                     action = successor.action
-                    actionCost = successor.actionCost
+                    actionCost = successor.actionCost.toLong()
                 }
 
                 logger.debug { "Expanding from $sourceNode --> $successorState :: open list size: ${openList.size}" }
@@ -279,7 +279,7 @@ class LssLrtaStarPlanner<StateType : State<StateType>>(val domain: Domain<StateT
             val undiscoveredNode = Node(
                     state = successorState,
                     heuristic = domain.heuristic(successorState),
-                    actionCost = successor.actionCost,
+                    actionCost = successor.actionCost.toLong(),
                     action = successor.action,
                     parent = parent,
                     cost = MAX_VALUE,
