+++ src/main/kotlin/edu/unh/cs/ai/realtimesearch/planner/realtime/SimpleSafe.kt
@@ -263,10 +263,10 @@ class SimpleSafePlanner<StateType : State<StateType>>(val domain: Domain<StateTy
             val successorGValueFromCurrent = currentGValue + successor.actionCost
             // always add the successor doing a BFS
             successorNode.apply {
-                cost = successorGValueFromCurrent
+                cost = successorGValueFromCurrent.toLong()
                 parent = sourceNode
                 action = successor.action
-                actionCost = successor.actionCost
+                actionCost = successor.actionCost.toLong()
                 depth = parent.depth + 1
             }
 
@@ -342,7 +342,7 @@ class SimpleSafePlanner<StateType : State<StateType>>(val domain: Domain<StateTy
                 successorNode.safe = true
             }
 
-            successorNode.predecessors.add(SearchEdge(node = sourceNode, action = successor.action, actionCost = successor.actionCost))
+            successorNode.predecessors.add(SearchEdge(node = sourceNode, action = successor.action, actionCost = successor.actionCost.toLong()))
 
             // out dated nodes are updated
             if (successorNode.iteration != iterationCounter) {
@@ -362,10 +362,10 @@ class SimpleSafePlanner<StateType : State<StateType>>(val domain: Domain<StateTy
             val successorGValueFromCurrent = currentGValue + successor.actionCost
             if (successorNode.cost > successorGValueFromCurrent) {
                 successorNode.apply {
-                    cost = successorGValueFromCurrent
+                    cost = successorGValueFromCurrent.toLong()
                     parent = sourceNode
                     action = successor.action
-                    actionCost = successor.actionCost
+                    actionCost = successor.actionCost.toLong()
                     depth = parent.depth + 1
                 }
 
@@ -400,7 +400,7 @@ class SimpleSafePlanner<StateType : State<StateType>>(val domain: Domain<StateTy
             val undiscoveredNode = Node(
                     state = successorState,
                     heuristic = domain.heuristic(successorState),
-                    actionCost = successor.actionCost,
+                    actionCost = successor.actionCost.toLong(),
                     action = successor.action,
                     parent = parent,
                     cost = MAX_VALUE,
